tooling/psql_python/requirements.txt
psycopg2-binary == 2.9.7